// src/Components/Auctions/NFTCard/AuctionCard.d.ts
export interface AuctionCardProps {
  item: { _id: string, cloudinaryUrl: string, startBid: number, tokenId: number, chain: number, lastBid: number, auctionId: number, nftId: string, name: string; royalty: string; highestBid: number, highestBidder: string, description: string, image: string, duration: number, views: number, sellerInfo: string, }
}